Created reuseable RangePanel component
