--- aim/web/ui/src/components/ImagesExploreRangePanel/index.ts
@@ -1,6 +1,0 @@
-import ImagesExploreRangePanel from './ImagesExploreRangePanel';
-
-export * from './types.d';
-export * from './ImagesExploreRangePanel';
-
-export default ImagesExploreRangePanel;
