--- aim/web/ui/src/components/ImagesExploreRangePanel/types.d.ts
@@ -1,13 +1,0 @@
-import { IMediaPanelProps } from '../MediaPanel';
-
-export interface IImagesExploreRangePanelProps {
-  recordSlice: number[];
-  indexSlice: number[];
-  indexRange: number[];
-  stepRange: number[];
-  indexDensity: number;
-  recordDensity: number;
-  applyButtonDisabled: boolean;
-  onSliceRangeChange: IMediaPanelProps['onSliceRangeChange'];
-  onDensityChange: IMediaPanelProps['onDensityChange'];
-}
