+++ aim/web/ui/src/components/RangePanel/index.tsx
@@ -1,0 +1,5 @@
+import RangePanel from './RangePanel';
+
+export * from './RangePanel.d';
+
+export default RangePanel;
