+++ aim/web/ui/src/components/kit/Icon/Icon.d.ts
@@ -129,4 +129,5 @@ export type IconName =
   | 'avatar'
   | 'branch'
   | 'color-scale-off'
-  | 'color-scale-on';
+  | 'color-scale-on'
+  | 'circle-info';
