+++ aim/web/ui/src/pages/RunDetail/AudiosVisualizer/AudiosVisualizer.scss
@@ -1,3 +1,3 @@
 .AudiosVisualizer {
-  height: calc(100% - 3.125rem);
+  height: calc(100% - 2.75rem);
 }
