+++ aim/web/ui/src/pages/RunDetail/ImagesVisualizer/ImagesVisualizer.scss
@@ -1,5 +1,5 @@
 @use 'src/styles/abstracts' as *;
 
 .ImagesVisualizer {
-  height: calc(100% - 3.125rem);
+  height: calc(100% - 2.75rem);
 }
