+++ aim/web/ui/src/pages/RunDetail/TraceVisualizationContainer/TraceVisualizationContainer.scss
@@ -25,41 +25,5 @@
     flex-direction: column;
     justify-content: space-between;
     background-color: $white;
-    .RangePanel {
-      display: flex;
-      align-items: center;
-      min-height: toRem(44px);
-      width: 100%;
-      justify-content: center;
-      background: $white;
-      border-top: $border-main;
-      border-bottom-right-radius: toRem(6px);
-      .RangePanelContainer {
-        display: flex;
-        align-items: center;
-        height: 100%;
-        width: 100%;
-        max-width: 85rem;
-        padding: 0 2.125rem;
-        justify-content: center;
-        .ApplyButtonContainer {
-          height: 100%;
-          display: flex;
-          align-items: center;
-          .ApplyButton {
-            height: 1.5rem !important;
-            width: 5.125rem;
-          }
-        }
-
-        .VerticalDivider {
-          width: 0.0625rem;
-          height: 1.5rem;
-          background: $primary-color-10;
-          margin-right: 2.75rem;
-          margin-left: 2.75rem;
-        }
-      }
-    }
   }
 }
